rename: change store folder name to stores
